refactor to use server actions for auth
