--- examples/with-supabase/app/auth/sign-out/route.ts
@@ -1,16 +1,0 @@
-import { createClient } from '@/utils/supabase/server'
-import { NextResponse } from 'next/server'
-import { cookies } from 'next/headers'
-
-export async function POST(request: Request) {
-  const requestUrl = new URL(request.url)
-  const cookieStore = cookies()
-  const supabase = createClient(cookieStore)
-
-  await supabase.auth.signOut()
-
-  return NextResponse.redirect(`${requestUrl.origin}/login`, {
-    // a 301 status is required to redirect from a POST to a GET route
-    status: 301,
-  })
-}
