+++ examples/with-supabase/components/AuthButton.tsx
@@ -1,6 +1,7 @@
 import { createClient } from '@/utils/supabase/server'
 import Link from 'next/link'
 import { cookies } from 'next/headers'
+import { redirect } from 'next/navigation'
 
 export default async function AuthButton() {
   const cookieStore = cookies()
@@ -10,10 +11,19 @@ export default async function AuthButton() {
     data: { user },
   } = await supabase.auth.getUser()
 
+  const signOut = async () => {
+    'use server'
+
+    const cookieStore = cookies()
+    const supabase = createClient(cookieStore)
+    await supabase.auth.signOut()
+    return redirect('/login')
+  }
+
   return user ? (
     <div className="flex items-center gap-4">
       Hey, {user.email}!
-      <form action="/auth/sign-out" method="post">
+      <form action={signOut}>
         <button className="py-2 px-4 rounded-md no-underline bg-btn-background hover:bg-btn-background-hover">
           Logout
         </button>
